requirements.txt
lxml==4.4.2
pandas==1.0.1
requests==2.21.0
PyYAML==5.3
